refactor message generation

+++ rosidl_typesupport_connext_cpp/rosidl_typesupport_connext_cpp/__init__.py
@@ -18,6 +18,7 @@ import os
 import re
 import subprocess
 
+from rosidl_cmake import convert_camel_case_to_lower_case_underscore
 from rosidl_generator_dds_idl import MSG_TYPE_TO_IDL
 from rosidl_parser import parse_message_file
 from rosidl_parser import parse_service_file
@@ -28,10 +29,10 @@ def parse_ros_interface_files(pkg_name, ros_interface_files):
     message_specs = []
     service_specs = []
     for idl_file in ros_interface_files:
-        filename, extension = os.path.splitext(idl_file)
+        extension = os.path.splitext(idl_file)[1]
         if extension == '.msg':
             message_spec = parse_message_file(pkg_name, idl_file)
-            message_specs.append(message_spec)
+            message_specs.append((idl_file, message_spec))
         elif extension == '.srv':
             service_spec = parse_service_file(pkg_name, idl_file)
             service_specs.append(service_spec)
@@ -40,11 +41,7 @@ def parse_ros_interface_files(pkg_name, ros_interface_files):
 
 def generate_dds_connext_cpp(
         pkg_name, dds_interface_files, dds_interface_base_path, deps,
-        output_dir, idl_pp, message_specs, service_specs):
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        pass
+        output_basepath, idl_pp, message_specs, service_specs):
 
     include_dirs = [dds_interface_base_path]
     for dep in deps:
@@ -58,11 +55,23 @@ def generate_dds_connext_cpp(
             include_dirs.append(idl_base_path)
 
     for index, idl_file in enumerate(dds_interface_files):
+        # get two level of parent folders for idl file
+        folder = os.path.dirname(idl_file)
+        parent_folder = os.path.dirname(folder)
+        output_path = os.path.join(
+            output_basepath,
+            os.path.basename(parent_folder),
+            os.path.basename(folder))
+        try:
+            os.makedirs(output_path)
+        except FileExistsError:
+            pass
+
         cmd = [idl_pp]
         for include_dir in include_dirs:
             cmd += ['-I', include_dir]
         cmd += [
-            '-d', output_dir,
+            '-d', output_path,
             '-language', 'C++',
             '-namespace',
             '-replace',
@@ -73,15 +82,15 @@ def generate_dds_connext_cpp(
 
         # modify generated code to support unbounded sequences and strings
         # http://community.rti.com/content/forum-topic/changing-max-size-sequence
-        unbounded_fields = [f for f in message_specs[index].fields
+        unbounded_fields = [f for f in message_specs[index][1].fields
                             if f.type.is_array and
                             f.type.array_size is None and
                             not f.type.is_upper_bound]
         if unbounded_fields:
             msg_name = os.path.splitext(os.path.basename(idl_file))[0]
-            msg_cxx_filename = os.path.join(output_dir, msg_name + '.cxx')
+            msg_cxx_filename = os.path.join(output_path, msg_name + '.cxx')
             _modify(msg_cxx_filename, unbounded_fields, _step_2_2)
-            plugin_cxx_filename = os.path.join(output_dir, msg_name + 'Plugin.cxx')
+            plugin_cxx_filename = os.path.join(output_path, msg_name + 'Plugin.cxx')
             _modify(plugin_cxx_filename, unbounded_fields, _step_2_1_and_2_3_and_2_4)
 
     return 0
@@ -250,40 +259,44 @@ def _get_dds_type(fields, field_name):
         idl_type = MSG_TYPE_TO_IDL[field.type.type]
         return IDL_TYPE_TO_DDS[idl_type]
     else:
-        return '%s::dds_::%s_' % (field.type.pkg_name, field.type.type)
+        return '%s::msg::dds_::%s_' % (field.type.pkg_name, field.type.type)
 
 
 def generate_cpp(
     pkg_name, message_specs, service_specs, known_msg_types, output_dir, template_dir
 ):
     mapping_msgs = {
-        os.path.join(template_dir, 'msg_TypeSupport.h.template'): '%s_TypeSupport.h',
-        os.path.join(template_dir, 'msg_TypeSupport.cpp.template'): '%s_TypeSupport.cpp',
+        os.path.join(template_dir, 'msg__type_support.hpp.template'): '%s__type_support.hpp',
+        os.path.join(template_dir, 'msg__type_support.cpp.template'): '%s__type_support.cpp',
     }
 
     mapping_srvs = {
-        os.path.join(template_dir, 'srv_ServiceTypeSupport.cpp.template'):
-        '%s_ServiceTypeSupport.cpp',
+        os.path.join(template_dir, 'srv__type_support.cpp.template'):
+        '%s__type_support.cpp',
     }
 
     for template_file in mapping_msgs.keys():
-        assert(os.path.exists(template_file))
+        assert os.path.exists(template_file), 'Could not find template: ' + template_file
 
     for template_file in mapping_srvs.keys():
-        assert(os.path.exists(template_file))
+        assert os.path.exists(template_file), 'Could not find template: ' + template_file
 
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        pass
+    functions = {
+        'get_header_filename_from_msg_name': convert_camel_case_to_lower_case_underscore,
+    }
 
-    for spec in message_specs:
+    for idl_file, spec in message_specs:
         validate_field_types(spec, known_msg_types)
+        subfolder = os.path.basename(os.path.dirname(idl_file))
         for template_file, generated_filename in mapping_msgs.items():
-            generated_file = os.path.join(output_dir, generated_filename % spec.base_type.type)
+            generated_file = os.path.join(
+                output_dir, subfolder, 'dds_connext', generated_filename %
+                convert_camel_case_to_lower_case_underscore(spec.base_type.type))
 
             try:
                 output = StringIO()
+                data = {'spec': spec, 'subfolder': subfolder}
+                data.update(functions)
                 # TODO reuse interpreter
                 interpreter = em.Interpreter(
                     output=output,
@@ -291,13 +304,14 @@ def generate_cpp(
                         em.RAW_OPT: True,
                         em.BUFFERED_OPT: True,
                     },
-                    globals={'spec': spec},
+                    globals=data,
                 )
                 interpreter.file(open(template_file))
                 content = output.getvalue()
                 interpreter.shutdown()
             except Exception:
-                os.remove(generated_file)
+                if os.path.exists(generated_file):
+                    os.remove(generated_file)
                 raise
 
             # only overwrite file if necessary
@@ -305,16 +319,24 @@ def generate_cpp(
                 with open(generated_file, 'r') as h:
                     if h.read() == content:
                         continue
+            try:
+                os.makedirs(os.path.dirname(generated_file))
+            except FileExistsError:
+                pass
             with open(generated_file, 'w') as h:
                 h.write(content)
 
     for spec in service_specs:
         validate_field_types(spec, known_msg_types)
         for template_file, generated_filename in mapping_srvs.items():
-            generated_file = os.path.join(output_dir, generated_filename % spec.srv_name)
+            generated_file = os.path.join(
+                output_dir, 'srv', 'dds_connext', generated_filename %
+                convert_camel_case_to_lower_case_underscore(spec.srv_name))
 
             try:
                 output = StringIO()
+                data = {'spec': spec}
+                data.update(functions)
                 # TODO reuse interpreter
                 interpreter = em.Interpreter(
                     output=output,
@@ -322,13 +344,14 @@ def generate_cpp(
                         em.RAW_OPT: True,
                         em.BUFFERED_OPT: True,
                     },
-                    globals={'spec': spec},
+                    globals=data,
                 )
                 interpreter.file(open(template_file))
                 content = output.getvalue()
                 interpreter.shutdown()
             except Exception:
-                os.remove(generated_file)
+                if os.path.exists(generated_file):
+                    os.remove(generated_file)
                 raise
 
             # only overwrite file if necessary
@@ -336,6 +359,10 @@ def generate_cpp(
                 with open(generated_file, 'r') as h:
                     if h.read() == content:
                         continue
+            try:
+                os.makedirs(os.path.dirname(generated_file))
+            except FileExistsError:
+                pass
             with open(generated_file, 'w') as h:
                 h.write(content)
